JSONFormatBear: Move to js
JSON is JavaScript Object Notation and thus fits into the JS section.
